+++ packages/expo-doctor/CHANGELOG.md
@@ -10,6 +10,14 @@
 
 ### 💡 Others
 
+- Inform users of corrupted `node_modules` if autolinking finds duplicates with all identical versions ([#39026](https://github.com/expo/expo/pull/39026) by [@kitten](https://github.com/kitten))
+
+## 1.16.0 — 2025-08-19
+
+### 🎉 New features
+
+- Add check for lock-files. ([#38963](https://github.com/expo/expo/pull/38963) by [@entiendonull](https://github.com/entiendonull))
+
 ## 1.15.1 — 2025-08-16
 
 _This version does not introduce any user-facing changes._